requirements.txt
mistralai
supabase
pdf2image
requests
pillow
python-dotenv
